requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-lru=2.0.4=pypi_0
attrs=24.3.0=pypi_0
babel=2.16.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
bleach=6.2.0=pypi_0
brotli-python=1.0.9=py312h6a678d5_9
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.12.31=h06a4308_0
catalyst=21.5=pypi_0
certifi=2024.12.14=pyhd8ed1ab_0
cffi=1.17.1=pypi_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=pypi_0
cuda-cudart=12.4.127=h99ab3db_0
cuda-cudart_linux-64=12.4.127=hd681fbe_0
cuda-cupti=12.4.127=h6a678d5_1
cuda-libraries=12.4.1=h06a4308_1
cuda-nvrtc=12.4.127=h99ab3db_1
cuda-nvtx=12.4.127=h6a678d5_1
cuda-opencl=12.4.127=h6a678d5_0
cuda-runtime=12.4.1=hb982923_0
cuda-version=12.4=hbda6634_3
cycler=0.12.1=pypi_0
debugpy=1.8.11=py312h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_1
defusedxml=0.7.1=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
expat=2.6.4=h6a678d5_0
fastjsonschema=2.21.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
fonttools=4.55.4=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.12.0=pypi_0
giflib=5.2.2=h5eee18b_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
h11=0.14.0=pypi_0
htmlmin=0.1.12=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.27.1=pypi_0
idna=3.7=py312h06a4308_0
imagehash=4.3.1=pypi_0
imageio=2.37.0=pypi_0
importlib-metadata=8.6.1=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=8.31.0=pyh707e725_0
ipywidgets=8.1.5=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=py312h06a4308_1
joblib=1.1.1=pypi_0
jpeg=9e=h5eee18b_3
json5=0.10.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter=1.1.1=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-events=0.11.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.15.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
jupyterlab=4.3.4=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=h143b758_0
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libcublas=12.4.5.8=h99ab3db_1
libcufft=11.2.1.3=h99ab3db_1
libcufile=1.9.1.3=h99ab3db_1
libcurand=10.3.5.147=h99ab3db_1
libcusolver=11.6.1.9=h99ab3db_1
libcusparse=12.3.1.170=h99ab3db_1
libdeflate=1.22=h5eee18b_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgomp=14.2.0=h77fa898_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.2.5.30=h99ab3db_1
libnvfatbin=12.4.127=h7934f7d_2
libnvjitlink=12.4.127=h99ab3db_1
libnvjpeg=12.3.1.117=h6a678d5_1
libpng=1.6.39=h5eee18b_0
libsodium=1.0.20=h4ab18f5_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py312h5eee18b_1
matplotlib=3.10.0=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
missingno=0.5.2=pypi_0
mistune=3.1.0=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_2
mkl_fft=1.3.11=py312h5eee18b_0
mkl_random=1.2.8=py312h526ad5a_0
mpmath=1.3.0=py312h06a4308_0
multimethod=2.0=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.5=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
nettle=3.7.3=hbbd107a_1
networkx=3.3=py312h06a4308_0
notebook=7.3.2=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
ocl-icd=2.3.2=h5eee18b_1
opencv-python=4.11.0.86=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.4.0=h7b32b05_1
overrides=7.7.0=pypi_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=pypi_0
pandas-profiling=3.2.0=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
phik=0.12.4=pypi_0
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.0.0=py312hcea889d_1
pip=24.2=py312h06a4308_0
platformdirs=4.3.6=pyhd8ed1ab_1
plotly=5.24.1=pypi_0
prometheus-client=0.21.1=pypi_0
prompt-toolkit=3.0.50=pyha770c72_0
protobuf=5.29.3=pypi_0
psutil=5.9.0=py312h5eee18b_1
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pycparser=2.22=pypi_0
pydantic=2.10.5=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.8=h5148396_0
python-dateutil=2.9.0.post0=pyhff2d567_1
python-json-logger=3.2.1=pypi_0
pytorch=2.5.1=py3.12_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pytz=2024.2=pypi_0
pywavelets=1.8.0=pypi_0
pyyaml=6.0.2=py312h5eee18b_0
pyzmq=26.2.0=py312h6a678d5_0
readline=8.2=h5eee18b_0
referencing=0.36.1=pypi_0
requests=2.32.3=py312h06a4308_1
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.22.3=pypi_0
safetensors=0.5.2=pypi_0
scikit-image=0.25.1=pypi_0
scipy=1.15.1=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.3=pypi_0
setuptools=75.1.0=py312h06a4308_0
six=1.17.0=pyhd8ed1ab_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.13.1=pypi_0
tangled-up-in-unicode=0.2.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tenacity=9.0.0=pypi_0
tensorboardx=2.6.2.2=pypi_0
terminado=0.18.1=pypi_0
tifffile=2025.1.10=pypi_0
timm=1.0.14=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h39e8969_0
torchtriton=3.1.0=py312
torchvision=0.20.1=py312_cu124
tornado=6.4.2=py312h5eee18b_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
types-python-dateutil=2.9.0.20241206=pypi_0
typing_extensions=4.12.2=py312h06a4308_0
tzdata=2025.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.3=py312h06a4308_0
visions=0.7.4=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.44.0=py312h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h3b0a872_7
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
